게시팔 CRUD 이미지 추가 완료

+++ React_SpringBoot/be/src/main/java/kr/or/ddit/be/mapper/BoardMapper.java
@@ -8,10 +8,10 @@ import java.util.List;
 @Mapper
 public interface BoardMapper {
     public List<BoardVO> list();
-    public BoardVO findById(Long id);
+    public BoardVO findById(int id);
 
     public int insert(BoardVO boardVO);
     public int update(BoardVO boardVO);
-    public int delete(Long boardId);
+    public int delete(int boardId);
 
 }
